// packages/wizzi.account/dist/src/server/features/config/types.ts
/*
    artifact generator: C:\My\wizzi\stfnbssl\wizzi\packages\wizzi-js\dist\lib\artifacts\ts\module\gen\main.js
    package: wizzi-js@0.7.8
    primary source IttfDocument: C:\My\wizzi\stfnbssl\wizzi\packages\wizzi.account\.wizzi\server\src\features\config\types.ts.ittf
    utc time: Tue, 25 May 2021 12:34:40 GMT
*/
export type ConfigType = { 
    port: number;
    sessionSecret: string;
    noCache: boolean;
    corsClientOrigin: string;
    mongoPath: string;
    mongoUser: string;
    mongoPassword: string;
    auth0Domain: string;
    auth0IssuerBaseURL: string;
    auth0ClientId: string;
    auth0BaseURL: string;
    auth0Secret: string;
    githubClientId: string;
    githubClientSecret: string;
    githubCallbackURL: string;
    isWizziDev: boolean;
    ittfPath: string;
    dataPath: string;
    metaHtmlIttfPath: string;
    metaFolderIttfPath: string;
    metaHtmlTextPath: string;
};